Two starter files added.
Both scripts are the initial templates that I will be using to create my project. 

The goal for the entire project is my attempt to learn how APT works as well as provide a detailed analysis of my thought process and hopefully this can serve as a template for future learners.